+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Fix deploy errors not being reported on first save after file creation
     * Fix "Always prompt for overwrite on save" issue #267
     * Fix "Setting org src as root causes `AuraDefinitionBundle metadata type not found in org`" issue #265
+    * Fix diffing returning expired access/refresh token #270
 * 3.15.2
     * Fix issue with first save of LWC or Aura components after creation. Users were being shown the file change notification.
     * Fix issue with hanging task after cancelling creating a file